server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
beautifulsoup4==4.12.2
gunicorn==21.2.0
pytest==7.4.2
black==23.9.1
pytest-cov==4.1.0
pymongo==4.5.0
python-dotenv==1.0.0
scikit-learn==1.4.0
numpy==1.26.4
matplotlib==3.8.3
requests==2.31.0
lxml==4.9.3 